requirements.txt
-i https://pypi.org/simple
aerich==0.7.2
aiosqlite==0.17.0 ; python_version >= '3.6'
annotated-types==0.7.0 ; python_version >= '3.8'
anyio==4.4.0 ; python_version >= '3.8'
appdirs==1.4.4
attrs==24.2.0 ; python_version >= '3.7'
black==24.8.0
certifi==2024.7.4 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
dictdiffer==0.9.0
idna==3.7 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
iso8601==1.1.0 ; python_version < '4.0' and python_full_version >= '3.6.2'
jedi==0.19.1 ; python_version >= '3.6'
mypy-extensions==1.0.0 ; python_version >= '3.5'
outcome==1.3.0.post0 ; python_version >= '3.7'
packaging==24.1 ; python_version >= '3.8'
parso==0.8.4 ; python_version >= '3.6'
pathspec==0.12.1 ; python_version >= '3.8'
platformdirs==4.2.2 ; python_version >= '3.8'
pluggy==1.5.0 ; python_version >= '3.8'
prompt-toolkit==3.0.47 ; python_full_version >= '3.7.0'
ptpython==3.0.29 ; python_version >= '3.7'
pydantic==2.8.2 ; python_version >= '3.8'
pydantic-core==2.20.1 ; python_version >= '3.8'
pygments==2.18.0 ; python_version >= '3.8'
pypika-tortoise==0.1.6 ; python_version >= '3.7' and python_version < '4.0'
pytest==8.3.2
pytz==2024.1
requests==2.32.3 ; python_version >= '3.8'
sniffio==1.3.1 ; python_version >= '3.7'
sortedcontainers==2.4.0
starlette==0.37.2 ; python_version >= '3.8'
starlette-testclient==0.4.1
tomlkit==0.13.0 ; python_version >= '3.8'
tortoise-cli==0.1.2
tortoise-orm[asyncmy]==0.21.5
trio==0.26.2
typing-extensions==4.12.2 ; python_version >= '3.8'
urllib3==2.2.2 ; python_version >= '3.8'
wcwidth==0.2.13
asyncmy==0.2.9
fastapi==0.112.0
h11==0.14.0 ; python_version >= '3.7'
httptools==0.6.1
python-dotenv==1.0.1
pyyaml==6.0.2
uvicorn[standard]==0.30.5
uvloop==0.19.0
watchfiles==0.23.0
websockets==12.0
